Abstract code & Remove redundant code & Simplfy code.
