+++ apm-sniffer/apm-sdk-plugin/httpClient-4.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/httpClient/v4/Constants.java
@@ -16,16 +16,13 @@
  *
  */
 
-package org.apache.skywalking.apm.plugin.httpClient.v4.define;
 
-import org.apache.skywalking.apm.agent.core.plugin.interceptor.ConstructorInterceptPoint;
-import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.ClassInstanceMethodsEnhancePluginDefine;
+package org.apache.skywalking.apm.plugin.httpClient.v4;
 
-public abstract class HttpClientInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
+/**
+ * @author Alan Lau
+ */
+public class Constants {
 
-    @Override
-    protected ConstructorInterceptPoint[] getConstructorsInterceptPoints() {
-        return null;
-    }
-
+    public static final String INTERCEPT_CLASS = "org.apache.skywalking.apm.plugin.httpClient.v4.HttpClientExecuteInterceptor";
 }
