+++ apm-sniffer/apm-sdk-plugin/httpClient-4.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/httpClient/v4/HttpClientExecuteInterceptor.java
@@ -21,6 +21,7 @@ package org.apache.skywalking.apm.plugin.httpClient.v4;
 import java.lang.reflect.Method;
 import java.net.MalformedURLException;
 import java.net.URL;
+
 import org.apache.http.HttpHost;
 import org.apache.http.HttpRequest;
 import org.apache.http.HttpResponse;
@@ -38,8 +39,9 @@ import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
 
 public class HttpClientExecuteInterceptor implements InstanceMethodsAroundInterceptor {
 
-    @Override public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments,
-                                       Class<?>[] argumentsTypes, MethodInterceptResult result) throws Throwable {
+    @Override
+    public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments,
+        Class<?>[] argumentsTypes, MethodInterceptResult result) throws Throwable {
         if (allArguments[0] == null || allArguments[1] == null) {
             // illegal args, can't trace. ignore.
             return;
@@ -51,12 +53,11 @@ public class HttpClientExecuteInterceptor implements InstanceMethodsAroundInterc
         String remotePeer = httpHost.getHostName() + ":" + port(httpHost);
 
         String uri = httpRequest.getRequestLine().getUri();
-        String requestURI = getRequestURI(uri);
-        String operationName = requestURI;
+        String operationName = getRequestURI(uri);
         AbstractSpan span = ContextManager.createExitSpan(operationName, contextCarrier, remotePeer);
 
         span.setComponent(ComponentsDefine.HTTPCLIENT);
-        Tags.URL.set(span, buildSpanValue(httpHost,uri));
+        Tags.URL.set(span, buildSpanValue(httpHost, uri));
         Tags.HTTP.METHOD.set(span, httpRequest.getRequestLine().getMethod());
         SpanLayer.asHttp(span);
 
@@ -67,8 +68,9 @@ public class HttpClientExecuteInterceptor implements InstanceMethodsAroundInterc
         }
     }
 
-    @Override public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments,
-                                        Class<?>[] argumentsTypes, Object ret) throws Throwable {
+    @Override
+    public Object afterMethod(EnhancedInstance objInst, Method method, Object[] allArguments,
+        Class<?>[] argumentsTypes, Object ret) throws Throwable {
         if (allArguments[0] == null || allArguments[1] == null) {
             return ret;
         }
@@ -90,8 +92,9 @@ public class HttpClientExecuteInterceptor implements InstanceMethodsAroundInterc
         return ret;
     }
 
-    @Override public void handleMethodException(EnhancedInstance objInst, Method method, Object[] allArguments,
-                                                Class<?>[] argumentsTypes, Throwable t) {
+    @Override
+    public void handleMethodException(EnhancedInstance objInst, Method method, Object[] allArguments,
+        Class<?>[] argumentsTypes, Throwable t) {
         AbstractSpan activeSpan = ContextManager.activeSpan();
         activeSpan.errorOccurred();
         activeSpan.log(t);
@@ -101,9 +104,9 @@ public class HttpClientExecuteInterceptor implements InstanceMethodsAroundInterc
         if (isUrl(uri)) {
             String requestPath = new URL(uri).getPath();
             return requestPath != null && requestPath.length() > 0 ? requestPath : "/";
-        } else {
-            return uri;
         }
+
+        return uri;
     }
 
     private boolean isUrl(String uri) {
@@ -112,18 +115,14 @@ public class HttpClientExecuteInterceptor implements InstanceMethodsAroundInterc
     }
 
     private String buildSpanValue(HttpHost httpHost, String uri) {
-        if (isUrl(uri)) {
-            return uri;
-        } else {
-            StringBuilder buff = new StringBuilder();
-            buff.append(httpHost.getSchemeName().toLowerCase());
-            buff.append("://");
-            buff.append(httpHost.getHostName());
-            buff.append(":");
-            buff.append(port(httpHost));
-            buff.append(uri);
-            return buff.toString();
-        }
+        return isUrl(uri) ? uri : new StringBuilder()
+            .append(httpHost.getSchemeName().toLowerCase())
+            .append("://")
+            .append(httpHost.getHostName())
+            .append(":")
+            .append(port(httpHost))
+            .append(uri).toString();
+
     }
 
     private int port(HttpHost httpHost) {
