+++ apm-sniffer/apm-sdk-plugin/httpClient-4.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/httpClient/v4/define/AbstractHttpClientInstrumentation.java
@@ -19,18 +19,15 @@
 
 package org.apache.skywalking.apm.plugin.httpClient.v4.define;
 
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.matcher.ElementMatcher;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.InstanceMethodsInterceptPoint;
 import org.apache.skywalking.apm.agent.core.plugin.match.ClassMatch;
+import org.apache.skywalking.apm.plugin.httpClient.v4.InstanceMehthodsInterceptPointImpl;
 
-import static net.bytebuddy.matcher.ElementMatchers.named;
 import static org.apache.skywalking.apm.agent.core.plugin.match.NameMatch.byName;
 
 /**
- * {@link AbstractHttpClientInstrumentation} presents that skywalking intercepts
- * AbstractHttpClient#doExecute
- * by using {@link HttpClientInstrumentation#INTERCEPT_CLASS}.
+ * {@link AbstractHttpClientInstrumentation} presents that skywalking intercepts AbstractHttpClient#doExecute by using
+ * {@link org.apache.skywalking.apm.plugin.httpClient.v4.Constants#INTERCEPT_CLASS}.
  *
  * @author zhangxin
  */
@@ -38,34 +35,22 @@ public class AbstractHttpClientInstrumentation extends HttpClientInstrumentation
 
     private static final String ENHANCE_CLASS = "org.apache.http.impl.client.AbstractHttpClient";
 
+    /**
+     * DefaultRequestDirector is default implement. usually use in version 4.0-4.2 since 4.3, this class is Deprecated.
+     */
     @Override
     public ClassMatch enhanceClass() {
         return byName(ENHANCE_CLASS);
     }
 
     /**
-     * version 4.2, intercept method: execute, intercept
-     * public final HttpResponse execute(HttpHost target, HttpRequest request, HttpContext context)
+     * version 4.2, intercept method: execute, intercept public final HttpResponse execute(HttpHost target, HttpRequest
+     * request, HttpContext context)
      */
     @Override
     protected InstanceMethodsInterceptPoint[] getInstanceMethodsInterceptPoints() {
         return new InstanceMethodsInterceptPoint[] {
-            new InstanceMethodsInterceptPoint() {
-                @Override
-                public ElementMatcher<MethodDescription> getMethodsMatcher() {
-                    return named("doExecute");
-                }
-
-                @Override
-                public String getMethodsInterceptor() {
-                    return getInstanceMethodsInterceptor();
-                }
-
-                @Override
-                public boolean isOverrideArgs() {
-                    return false;
-                }
-            }
+            new InstanceMehthodsInterceptPointImpl()
         };
     }
 }
